能够正常编码输出packet，同时构建了构建者模式和代理模式，项目极大解耦

+++ app/src/main/cpp/src/utils/XData.h
@@ -15,7 +15,7 @@ enum
 
 struct XData {
     int type = 0;
-    int pts = 0;
+    int64_t pts = 0;
     u_char *data = 0;
     int size = 0;
     bool isAudio = false;
